requirements.txt
# Pinned for reproducible builds; update periodically
fastapi>=0.109,<0.115
pydantic>=2.0,<3
uvicorn>=0.27,<0.32
requests>=2.31,<3
httpx>=0.26,<0.28
python-dotenv>=1.0,<1.1
PyPDF2>=3.0,<4
python-docx>=1.0,<1.2
beautifulsoup4>=4.12.3,<4.13
qrcode[pil]>=7.4,<8

# LangChain
langchain>=0.2.17,<0.3
langchain-openai>=0.1,<0.2
langchain-qdrant>=0.1,<0.2
langchain-community>=0.2,<0.3
langchain-core>=0.2,<0.3

# Vector DB
qdrant-client>=1.7,<1.12

# Optional: docling (document conversion)
docling>=2.0,<3

# Celery (background ingestion)
celery[redis]>=5.3,<6

# Testing
pytest>=7.4,<9
